fix: columna activo en tablas maestras, tabla seguimientos, menu duplicado y exportar excel
- ALTER TABLE ejecutado en executives, directores, lineas y codigos
- Corregido toggle activo en modulos maestros (EjecutivosCrud, CodigosClienteCrud)
- Eliminado enlace duplicado de seguimiento en layouts/app.blade.php
- Creadas tablas facturas y auditoria_observaciones
- Corregidos errores en SeguimientoForm (else huerfano, variable isAdmin, relacion auditorias)
- Agregado SeguimientosExport y boton exportar Excel en vista de seguimientos
- Seeder SeguimientosSeeder con 7 registros de prueba (todos los estados del enum)

+++ database/migrations/2026_05_11_000001_create_seguimientos_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('seguimientos', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('marca_id')->nullable();
+            $table->unsignedInteger('marca_id')->nullable();
             $table->foreign('marca_id')->references('id')->on('marcas')->nullOnDelete();
             $table->string('cliente');
             $table->string('linea_primaria')->nullable();
